new init.

+++ WXDemo/proto/Mm.pbobjc.h
@@ -2606,7 +2606,7 @@ typedef GPB_ENUM(BaseRequest_FieldNumber) {
 /** Test to see if @c sessionKey has been set. */
 @property(nonatomic, readwrite) BOOL hasSessionKey;
 
-@property(nonatomic, readwrite) int32_t uin;
+@property(nonatomic, readwrite) int64_t uin;
 
 @property(nonatomic, readwrite) BOOL hasUin;
 /** 手机guid长度16,这里取前15字节,以'\\0'结尾 */
@@ -3225,7 +3225,7 @@ typedef GPB_ENUM(ManualAuthResponse_AuthParam_FieldNumber) {
 
 @interface ManualAuthResponse_AuthParam : GPBMessage
 
-@property(nonatomic, readwrite) int32_t uin;
+@property(nonatomic, readwrite) int64_t uin;
 
 @property(nonatomic, readwrite) BOOL hasUin;
 @property(nonatomic, readwrite, strong, null_resettable) ManualAuthResponse_AuthParam_Ecdh *ecdh;
